move sort array helper method to ArrayUtils incl. renaming test class (ArrayUtilsTest), to match class under test name
